auth-user/requirements.txt
psycopg2-binary
bcrypt
pyotp
cryptography